Move EngineProtocol to SyncDependencies
